+++ src/base/classes/CustomClient.ts
@@ -27,7 +27,11 @@ export default class CustomClient extends Client implements ICustomClient {
 
   constructor() {
     super({
-      intents: [GatewayIntentBits.Guilds],
+      intents: [
+        GatewayIntentBits.Guilds,
+        GatewayIntentBits.GuildMessages,
+        GatewayIntentBits.MessageContent,
+      ],
       partials: [Partials.Channel],
     });
     this.config = config;